refactor: Move files to same folder structure
Moved user-related functions from 'firebaseUserActions' to 'user' directory, deleted 'firebaseUserActions' directory, and updated imports accordingly.
